requirements.txt
Flask==2.2.3
requests==2.28.2
yoyo-migrations==8.2.0
filetype==1.2.0
pillow==9.5.0
